feat: refactor hooks
